backend/requirements.txt
fastapi==0.112.0
fuzzywuzzy==0.18.0
langchain==0.2.12
langchain_core==0.2.29
langchain_openai==0.1.20
langsmith==0.1.98
pydantic==2.8.2
pyprojroot==0.3.0
python-dotenv==1.0.1
PyYAML==6.0.2
Requests==2.32.3
uvicorn==0.30.5
dotenv